modul warehouse (done),modul periode(done),modul measure(done),modul product(done),update modul transaction sell/buy/stock opname(done), response failed to login(done)

+++ src/main/java/com/backend/inventaris/model/Transaction.java
@@ -23,7 +23,7 @@ public class Transaction {
     private Long price;
 
     @Column(name = "qty")
-    private Integer qty;
+    private Long qty;
 
     @Enumerated(EnumType.STRING)
     private TypeTransaction typeTransaction;
@@ -80,11 +80,11 @@ public class Transaction {
         this.price = price;
     }
 
-    public Integer getQty() {
+    public Long getQty() {
         return qty;
     }
 
-    public void setQty(Integer qty) {
+    public void setQty(Long qty) {
         this.qty = qty;
     }
